исправил ошибки подфазы востановления в фазе атаки

+++ src/mechanics/db/localGame/update/updateEffects.go
@@ -8,40 +8,43 @@ import (
 )
 
 func UnitEffects(unit *unit.Unit) error {
-	for _, unitEffect := range unit.Effects {
-		if unitEffect.StepsTime == 0 {
+	for i, unitEffect := range unit.Effects {
+		if unitEffect != nil {
+			if unitEffect.StepsTime == 0 {
 
-			_, err := dbConnect.GetDBConnect().Exec("DELETE FROM action_game_unit_effects WHERE id=$1", unitEffect.ID)
-
-			if err != nil {
-				println("Error delete unit effect")
-				log.Fatal(err)
-				return err
-			}
-
-		} else {
-			if unitEffect.ID != 0 {
-
-				_, err := dbConnect.GetDBConnect().Exec("UPDATE action_game_unit_effects SET left_steps=$1, id_effect=$3 WHERE id=$2", unitEffect.StepsTime, unitEffect.ID, unitEffect.TypeID)
+				_, err := dbConnect.GetDBConnect().Exec("DELETE FROM action_game_unit_effects WHERE id=$1", unitEffect.ID)
 
 				if err != nil {
-					println("Error update unit effect")
+					println("Error delete unit effect")
 					log.Fatal(err)
 					return err
 				}
 
+				unit.Effects[i] = nil // удаляем эфект у юнита
 			} else {
-				id := 0
-				err := dbConnect.GetDBConnect().QueryRow("INSERT INTO action_game_unit_effects (id_unit, id_effect, left_steps) "+
-					"VALUES ($1, $2, $3) RETURNING id", unit.ID, unitEffect.TypeID, unitEffect.StepsTime).Scan(&id)
+				if unitEffect.ID != 0 {
 
-				if err != nil {
-					println("Error add new unit effect")
-					log.Fatal(err)
-					return err
+					_, err := dbConnect.GetDBConnect().Exec("UPDATE action_game_unit_effects SET left_steps=$1, id_effect=$3 WHERE id=$2", unitEffect.StepsTime, unitEffect.ID, unitEffect.TypeID)
+
+					if err != nil {
+						println("Error update unit effect")
+						log.Fatal(err)
+						return err
+					}
+
+				} else {
+					id := 0
+					err := dbConnect.GetDBConnect().QueryRow("INSERT INTO action_game_unit_effects (id_unit, id_effect, left_steps) "+
+						"VALUES ($1, $2, $3) RETURNING id", unit.ID, unitEffect.TypeID, unitEffect.StepsTime).Scan(&id)
+
+					if err != nil {
+						println("Error add new unit effect")
+						log.Fatal(err)
+						return err
+					}
+
+					unitEffect.ID = id
 				}
-
-				unitEffect.ID = id
 			}
 		}
 	}
